checkpoint(rust): WIP Go→Rust rewrite — cf-gojson keystone + cf-anomaly reconstructed
Safety checkpoint for the in-progress Rust rewrite (rust/ was untracked).

Verified green this session:
- cf-gojson keystone: GoValue/GoMap/MapOrigin + Go encoding/json byte-parity
  marshal + Encoder builder + Go shortest-float ftoa (19/19 tests).
- cf-anomaly: fully reconstructed from Go after concurrent-workflow corruption
  (statistics/zscore/detect/model; 9/9 tests).
- cf-version, cf-gojson tests: de-corrupted.

Known RED (see specs/rust-rewrite/STATUS.md): cf-analyze (~21 errors),
cf-couples + cf-uast-node corruption, cf-goyaml scaffold, CLI binaries unbuilt,
0/7 byte-identity captures verified.

+++ rust/crates/cf-streaming/tests/planner.rs
@@ -136,7 +136,7 @@ fn planner_low_growth_rate_hits_max_cap() {
     };
     let chunks = p.plan();
     let chunk_size = chunks[0].end - chunks[0].start;
-    assert_eq!(chunk_size, MAX_CHUNK_SIZE);
+    assert_eq!(chunk_size as i64, MAX_CHUNK_SIZE);
 }
 
 // --- PlanFrom ---
